Pre-index chat viewer IDs when cached news changes
- Added `cachedSharedViewInIds` map to `ChatHistoryAdapter`.
- Extracted JSON parsing logic into a helper method `extractSharedViewInIds` to parse viewer IDs dynamically from the `newsList`.
- Replaced the linear filter and JSON parse logic inside `getSharedViewInIds` with an O(1) map lookup.
- Handled malformed JSON gracefully by returning an empty list for that entry.
- Added `ChatHistoryAdapterTest` covering the duplicate merging, missing ID caching, and malformed JSON robustness of the `extractSharedViewInIds` extraction logic.
